requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
numpy==2.2.6
pandas==2.2.3
ta==0.11.0
requests==2.31.0
apscheduler==3.10.4
anthropic==0.40.0
eventlet==0.35.1
gunicorn==21.2.0
python-dotenv==1.0.0
